refactor(lua): Move Lua module specific code to isolated file
